+++ aws/fetch/gen_fetchers.go
@@ -1545,3 +1545,9 @@ func BuildCodepipelineFetchFuncs(conf *Config) fetch.Funcs {
 	}
 	return funcs
 }
+func BuildCodebuildFetchFuncs(conf *Config) fetch.Funcs {
+	funcs := make(map[string]fetch.Func)
+
+	addManualCodebuildFetchFuncs(conf, funcs)
+	return funcs
+}