+++ aws/services/gen_services.go
@@ -43,6 +43,8 @@ import (
 	cloudwatchtypes "github.com/aws/aws-sdk-go-v2/service/cloudwatch/types"
 	cloudwatchlogs "github.com/aws/aws-sdk-go-v2/service/cloudwatchlogs"
 	cloudwatchlogstypes "github.com/aws/aws-sdk-go-v2/service/cloudwatchlogs/types"
+	codebuild "github.com/aws/aws-sdk-go-v2/service/codebuild"
+	codebuildtypes "github.com/aws/aws-sdk-go-v2/service/codebuild/types"
 	codepipeline "github.com/aws/aws-sdk-go-v2/service/codepipeline"
 	codepipelinetypes "github.com/aws/aws-sdk-go-v2/service/codepipeline/types"
 	configservice "github.com/aws/aws-sdk-go-v2/service/configservice"
@@ -133,6 +135,7 @@ var ServiceNames = []string{
 	"kinesis",
 	"redshift",
 	"codepipeline",
+	"codebuild",
 }
 
 var ResourceTypes = []string{
@@ -212,6 +215,7 @@ var ResourceTypes = []string{
 	"redshiftcluster",
 	"redshiftsubnetgroup",
 	"pipeline",
+	"buildproject",
 }
 
 var ServicePerAPI = map[string]string{
@@ -251,6 +255,7 @@ var ServicePerAPI = map[string]string{
 	"kinesis":                "kinesis",
 	"redshift":               "redshift",
 	"codepipeline":           "codepipeline",
+	"codebuild":              "codebuild",
 }
 
 var ServicePerResourceType = map[string]string{
@@ -330,6 +335,7 @@ var ServicePerResourceType = map[string]string{
 	"redshiftcluster":     "redshift",
 	"redshiftsubnetgroup": "redshift",
 	"pipeline":            "codepipeline",
+	"buildproject":        "codebuild",
 }
 
 var APIPerResourceType = map[string]string{
@@ -409,6 +415,7 @@ var APIPerResourceType = map[string]string{
 	"redshiftcluster":     "redshift",
 	"redshiftsubnetgroup": "redshift",
 	"pipeline":            "codepipeline",
+	"buildproject":        "codebuild",
 }
 
 type Infra struct {
@@ -4826,4 +4833,132 @@ func (s *Codepipeline) FetchByType(ctx context.Context, t string) (cloud.GraphAP
 
 func (s *Codepipeline) IsSyncDisabled() bool {
 	return !getBool(s.config, "aws.codepipeline.sync", true)
+}
+
+type Codebuild struct {
+	fetcher         fetch.Fetcher
+	region, profile string
+	config          map[string]any
+	log             *logger.Logger
+	CodebuildClient *codebuild.Client
+}
+
+func NewCodebuild(cfg aws.Config, profile string, extraConf map[string]any, log *logger.Logger) cloud.Service {
+	region := cfg.Region
+	codebuildClient := codebuild.NewFromConfig(cfg)
+
+	fetchConfig := awsfetch.NewConfig(
+		codebuildClient,
+	)
+	fetchConfig.Extra = extraConf
+	fetchConfig.Log = log
+
+	return &Codebuild{
+		CodebuildClient: codebuildClient,
+		fetcher:         fetch.NewFetcher(awsfetch.BuildCodebuildFetchFuncs(fetchConfig)),
+		config:          extraConf,
+		region:          region,
+		profile:         profile,
+		log:             log,
+	}
+}
+
+func (s *Codebuild) Name() string {
+	return "codebuild"
+}
+
+func (s *Codebuild) Region() string {
+	return s.region
+}
+
+func (s *Codebuild) Profile() string {
+	return s.profile
+}
+
+func (s *Codebuild) ResourceTypes() []string {
+	return []string{
+		"buildproject",
+	}
+}
+
+func (s *Codebuild) Fetch(ctx context.Context) (cloud.GraphAPI, error) {
+	if s.IsSyncDisabled() {
+		return graph.NewGraph(), nil
+	}
+
+	allErrors := new(fetch.Error)
+
+	gph, err := s.fetcher.Fetch(context.WithValue(ctx, "region", s.region))
+	defer s.fetcher.Reset()
+
+	for _, e := range *fetch.WrapError(err) {
+		switch ee := e.(type) {
+		case nil:
+			continue
+		default:
+			var ae smithy.APIError
+			if errors.As(ee, &ae) && ae.ErrorMessage() == accessDenied {
+				allErrors.Add(cloud.ErrFetchAccessDenied)
+			} else {
+				allErrors.Add(ee)
+			}
+		}
+	}
+
+	if err := gph.AddResource(graph.InitResource(cloud.Region, s.region)); err != nil {
+		return gph, err
+	}
+
+	snap := gph.AsRDFGraphSnaphot()
+
+	errc := make(chan error)
+	var wg sync.WaitGroup
+	if getBool(s.config, "aws.codebuild.buildproject.sync", true) {
+		list, err := s.fetcher.Get("buildproject_objects")
+		if err != nil {
+			return gph, err
+		}
+		if _, ok := list.([]codebuildtypes.Project); !ok {
+			return gph, errors.New("cannot cast to '[]codebuildtypes.Project' type from fetch context")
+		}
+		for _, r := range list.([]codebuildtypes.Project) {
+			for _, fn := range addParentsFns["buildproject"] {
+				wg.Add(1)
+				go func(f addParentFn, snap tstore.RDFGraph, region string, res *codebuildtypes.Project) {
+					defer wg.Done()
+					err := f(gph, snap, region, res)
+					if err != nil {
+						errc <- err
+						return
+					}
+				}(fn, snap, s.region, &r)
+			}
+		}
+	}
+
+	go func() {
+		wg.Wait()
+		close(errc)
+	}()
+
+	for err := range errc {
+		if err != nil {
+			allErrors.Add(err)
+		}
+	}
+
+	if allErrors.Any() {
+		return gph, allErrors
+	}
+
+	return gph, nil
+}
+
+func (s *Codebuild) FetchByType(ctx context.Context, t string) (cloud.GraphAPI, error) {
+	defer s.fetcher.Reset()
+	return s.fetcher.FetchByType(context.WithValue(ctx, "region", s.region), t)
+}
+
+func (s *Codebuild) IsSyncDisabled() bool {
+	return !getBool(s.config, "aws.codebuild.sync", true)
 }