+++ template/internal/ast/gen_entities.go
@@ -30,6 +30,7 @@ var entities = map[Entity]struct{}{
 	"appscalingpolicy":    {},
 	"appscalingtarget":    {},
 	"bucket":              {},
+	"buildproject":        {},
 	"cachecluster":        {},
 	"cachesubnetgroup":    {},
 	"certificate":         {},